cafe_service/requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
cryptography==40.0.1
Django==3.1.14
django-haystack==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
docopt==0.6.2
drf-haystack==1.8.11
drf-yasg==1.21.5
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-watch==4.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
watchdog==3.0.0
whitenoise==6.4.0
Whoosh==2.7.4
woosh==0.0.1
yml==0.0.1
